// src/app/shared/api/model/placeAdminDTO.ts
/**
 * Generic Porta;
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface PlaceAdminDTO {
    placeId?: string;
    icon?: string;
    name?: string;
    rating?: string;
    vicinity?: string;
    address?: string;
    categoriesList?: string;
    imageUrl?: string;
    customName?: string;
    customRating?: string;
    customDescription?: string;
    customRecommended?: boolean;
    customImageUrl?: string;
    customIcon?: string;
    customCaregoriesList?: string;
    longitude?: number;
    latitude?: number;
    distance?: number;
    id?: string;
}